feat: renamed camscroll to parallax component
