+++ .gitignore
@@ -11,3 +11,4 @@ replit.nix
 .pnpm-store
 tsconfig.vitest-temp.json
 src/index.ts
+bun.lock
