+++ package.json
@@ -116,6 +116,11 @@
       "vite@<=5.4.19": ">=5.4.20",
       "axios@>=1.0.0 <1.12.0": ">=1.12.0",
       "tar-fs@>=3.0.0 <3.1.1": ">=3.1.1"
-    }
+    },
+    "onlyBuiltDependencies": [
+      "dprint",
+      "esbuild",
+      "puppeteer"
+    ]
   }
 }
